--- a/test/unit-tests/utils/function.test.js
+++ b/test/unit-tests/utils/function.test.js
@@ -31,8 +31,8 @@ describe('util.function', function () {
       const m = memoize(f)
 
       assert.strictEqual(m({ x: 2, y: 3 }), 6)
-      assert.deepStrictEqual(Object.keys(m.cache), ['[{"x":2,"y":3}]'])
-      assert.strictEqual(m.cache['[{"x":2,"y":3}]'], 6)
+      assert.deepStrictEqual(Object.keys(m.cache.values), ['[{"x":2,"y":3}]'])
+      assert.strictEqual(m.cache.values['[{"x":2,"y":3}]'], 6)
     })
 
     it('should memoize a function with a custom hashIt function', function () {
@@ -41,11 +41,11 @@ describe('util.function', function () {
         return 'id:' + args[0].id
       }
 
-      const m = memoize(f, hashIt)
+      const m = memoize(f, { hasher: hashIt })
 
       assert.strictEqual(m({ id: 2 }), 2)
-      assert.deepStrictEqual(Object.keys(m.cache), ['id:2'])
-      assert.strictEqual(m.cache['id:2'], 2)
+      assert.deepStrictEqual(Object.keys(m.cache.values), ['id:2'])
+      assert.strictEqual(m.cache.values['id:2'], 2)
     })
 
     it('should really return the cached result', function () {
@@ -58,6 +58,20 @@ describe('util.function', function () {
       a = 3
       assert.strictEqual(m(4), 2)
     })
+  })
+
+  it('should limit the number of values stored', function () {
+    let a = 1
+    const f = function (x) { a++; return a } // trick: no pure function
+
+    const m = memoize(f, { limit: 2 })
+
+    assert.strictEqual(m(1), 2)
+    assert.strictEqual(m(2), 3)
+    // this should evict m(1)
+    assert.strictEqual(m(3), 4)
+
+    assert.strictEqual(m(1), 5)
   })
 
   describe('memoizeCompare', function () {