@josdejong 2020-12-23
Thanks Flaviu! Your PR looks good at first sight 👍  nice idea to add memoization to `_findUnit`.

Some feedbacks:

1. you're talking about performance improvements. Do you have any feel of how much faster this solution is in practice? (there is a `benchmark` folder where you could add a small benchmark to try out). Since we're adding code, it should be worth it :)
2. It is possible to create new units with `createUnit`. I think that the cached results of `_findUnit` should be reset when there is a new unit created.
3. I see a unit test is still failing, can you look into that?