@josdejong 2020-12-24
wow, that is a huge performance improvement! That's definitely worth it 😃 👍 , very nice!

Thanks for your last updates Flaviu, will now merge your PR.