@josdejong 2020-12-30
This improvement is published now in `v8.1.1`. Thanks again!